[ oorexx-Bugs-3534402 ] do over not working with strings
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@7875 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ interpreter/execution/RexxNativeActivation.cpp
@@ -1441,7 +1441,7 @@ void RexxNativeActivation::callRegisteredRoutine(RoutineClass *_routine, Registe
         enableVariablepool();                // enable the variable pool interface here
         activity->releaseAccess();           /* force this to "safe" mode         */
         // now process the function call
-        functionrc = (*methp)(functionName->getStringData(), count, argPtr, queuename, &funcresult);
+        functionrc = (int)(*methp)(functionName->getStringData(), count, argPtr, queuename, &funcresult);
         activity->requestAccess();           /* now in unsafe mode again          */
     }
     catch (RexxActivation *)
